+++ examples/esbuild-pub/dubhe.config.ts
@@ -6,7 +6,7 @@ export default {
     hello: './src/HelloWorld.vue',
     module: './src/module.ts',
   },
-  types: false,
+  types: true,
   // limit: 1000,
   externals: (id) => {
     if (id === 'vue')