+++ packages/pub/src/esbuild.ts
@@ -6,9 +6,13 @@ import { BundlePlugin as Bundle, CSSPlugin as CSS } from './bundle/esbuild'
 export { dtsPlugin as Dts }
 export { CSS, Bundle }
 
-export function Pub(dubheConfig: PubConfig): any {
-  const config = Object.assign({}, dubheConfig) as Required<PubConfig>
+export function Pub(config: PubConfig) {
+  const plugins: any = []
+  config.types && plugins.push(dtsPlugin(config).esbuild)
+  plugins.push(Bundle(config))
 
-  return [dtsPlugin(config).esbuild, Bundle(config), CSS()]
+  plugins.push(CSS())
+
+  return plugins
 }