+++ packages/lib/CHANGELOG.md
@@ -1,5 +1,15 @@
 # dubhe
 
+## 2.0.0
+
+### Major Changes
+
+- 756ffad: add support for app&lib(only in vite);add hash support;refactor virtual entry; refactor importmap format
+
+### Patch Changes
+
+- a5b2537: cli can handle hash correctly and download file on demand
+
 ## 2.0.0-alpha.1
 
 ### Patch Changes